:recycle: Move common mocks to conftest
