+++ tests/projects/test_cavatica_sync.py
@@ -1,15 +1,14 @@
 import pytest
 import pytz
 from unittest.mock import MagicMock
-from dataclasses import dataclass
 from datetime import datetime
-from typing import Callable
 from creator.projects.cavatica import (
     sync_cavatica_projects,
     sync_cavatica_account,
 )
 from creator.studies.models import Study
 from creator.projects.models import Project
+from conftest import CavaticaProject
 
 
 SYNC_PROJECTS_MUTATION = """
@@ -42,44 +41,6 @@ def mock_sync_cavatica_account(mocker):
     )
     sync_cavatica_account.return_value = [], [], []
     return sync_cavatica_account
-
-
-@dataclass
-class CavaticaProject:
-    id: str = "author/test-harmonization"
-    name: str = "Test name-bwa-mem"
-    description: str = "Test description"
-    href: str = "test_url"
-    created_by: str = "author"
-    created_on: datetime = datetime(
-        2018, 8, 13, 15, 16, 17, 0, tzinfo=pytz.utc
-    )
-    modified_on: datetime = datetime(
-        2018, 10, 13, 15, 16, 17, 0, tzinfo=pytz.utc
-    )
-    save: Callable = MagicMock()
-
-
-@pytest.fixture
-def mock_cavatica_api(mocker):
-    """ Mocks out api for creating Cavatica project functions """
-
-    sbg = mocker.patch("creator.projects.cavatica.sbg")
-
-    # Project subresource of the sbg api
-    ProjectMock = MagicMock()
-    project_list = [
-        CavaticaProject(id="test_id_01_harmonization"),
-        CavaticaProject(id="test_id_02_harmonization"),
-    ]
-    ProjectMock.query().all.return_value = project_list
-
-    Api = MagicMock()
-    Api().projects = ProjectMock
-
-    sbg.Api = Api
-
-    return sbg
 
 
 def test_correct_sync(db, mock_sync_cavatica_account):
